Add ESLint rules
This adds some relatively basic ESLint rules for the project. Probably
the most controversial in all of this is using a default of 2 spaces for
tabs, which is just shy of heresy. Probably the only thing more heretical
would be to mix tabs and spaces, but which project would ever do something
as insane as that?