lib/errors: Add default HTTP errors to force JSON responses
By default, Express will not send errors back as JSON. This can
result in some reptetive boilerplate leaking around, as well as
making it easy to use 'sendStatus' which will not produce the result
we want. While it's a bit funky, this adds a set of Errors that can
be created and passed to next(), which will percolate down to the
default error handler. The default error handler in turn will properly
send the HTTP error as JSON.