controllers/login: Add a login route
This adds the ability to log into an account. Upon logging in,
a Session is created and a token returned from it that should
be used for subsequent auth requests.

At some point in the future we may want to switch this to JWT
to take advantage of authorization, but for now this suffices
for our purposes.