Look: Add the Look command
This adds the Look command, mostly ported over from the failed
attempt at making this game (haha... ugh.) This consists of two
parts: A factory which will generate the action from the supplied
tokens, and the action itself which takes the player character
and manipulates them and their environment. The test covers
looking at both the current room as well as those rooms that may
or may not be adjacent.

Note that eventually you can also look at objects and other things,
which is comented out for now. We'll get there.