CommandSet: Add the 'default' command set
Commands come in bundles based on some loose principles of what you
are allowed to do at a particular time. This can be based on authorization
or the environment that you're in. For now, we are creating a 'default'
command set - which are the things that players are allowed to do most
of the time. This only contains 'Look' for now, as well, that's our only
existing command.