room: Add a toText function that provides more details
When a character interacts with a room, they need to be able
to see what's around them. While this is a work in progress -
we'll eventually want a lot more interesting things to happen
when a character looks around - this adds a bit of text and
let's the character know what exits are around them.