playerCharacter: Wire through transports, to message parsing, to commands
This ties all the various pieces together. When a new connection is
made, a transport is set on the character. Setting a transport on the
character creates a listener callback for new messages. When a new message
is received, we parse it into JSON and look through the player character's
command sets to find one that can handle the message. When we have a command
action, we execute it, passing it the player character so it can perform the
action on the character and/or their environment.