world: Add an event loop
This patch adds an event loop that fires a configurable amount of time,
with a default of 3 seconds. This also adds a `shutdown` method to
the world. Because we now want to shutdown all the event loops that
will start firing, we add a shutdown method and invoke it from our
existing unit tests.