game/objects: Add Weapons
This adds a basic Weapon class that can express most any weapon
a character may want. For some ease of use, factory methods have
been added to generate a few basic objects.

We may or may not keep that. Not sure yet. It may also be better
down in the model, as we typically instantiate Weapon objects
on the fly.