game/objects: Add Armor
This adds 'armor' as a thing in the game. It can be worn, but it
can also be a container for other things. This gives us some
flexibility in what all we think about as 'armor': a piece of
clothing, a bracer, or a backpack. Could be a potion belt. Etc.

Some additional basic properties/checks have been put in here that
should likely get replicated out to weapons as well, such as checking
to see if a player can use the piece of armor. We'll want a similar
method with a similar method call. Likewise, being able to look at
what a thing 'is' via a type field is helpful when reversing the
process and saving the item. We should add that too.