db/models/Character: Allow Characters to have inanimates
While we could force a character to keep track of everything
by associating all items with locations, this is kind of a pain
and it's far easier to allow a character to 'haul' some things.
This adds a slot for character to do just that, which even if not
realistic, is far more kind to people.