playerCharacter: Add hauling items
This patch adds the ability for a character to haul an item around, like
bad luck. Or something.

As they haul items, they need to be able to have their weight increase.
So this adds a 'carryWeight' attribute that is modified as they carry
more stuff. If it exceed a maximum carry weight, they get encumbered and
start to burn more energy to move. This increases as they carry even more
crap.

So don't haul too much junk in that trunk.

(snort)