armor: Have a callback in case this container is on a person
When someone wants an item, they can get it from a room or from
an existing container. When they get it from an existing container,
they can get it from a storage container that is currently on a person.
That needs to therefore adjust the weight, as it is coming out of
a potentially weight reducing container and into the person's holding
area o' junk. The callback gives us an opportunity to adjust weights
when an item is added or removed from a container.