commands/(Get|Drop)Item: Advertise when a player gets or drops an item
Part of being in a world is seeing when other players do things. This
commit updates the Get/Drop Item commands to broadcast to a room when
a player gets or drops an item. As a consequence of this, we've now
added a small command to rooms to send messages directly to everyone in
a room from a particular player.

It's pretty obvious that having a Javascript array for storing items
in rooms/players/armor/etc. isn't going to work out long term. Too much
business logic gets splattered around in the commands. We'll need to
refactor this soon-ish, particularly once we need to handle manipulating
multiple objects with the same name. (Not to mention case sensitivity is
going to be a problem soon.)