Look: Add support for looking at objects
This patch lets us look at objects that are sitting around in a room.
We now return a longer description of the object. This makes sense
for looking *at* things, which gives us a bit of information that is
around us.

Eventually we'll add 'examine' which will be focused on things you
have in your possession.