loadCharacter: Prep for some additional characer types
We know we're going to have more than animals and humans, but we
have to start somewhere. This patch creates a loading function that
creates the objects from saved models. We'll need to do some heavy
refactoring eventually... but I'm not sure yet what should be in a
PlayerCharacter versus a human.

On the plus side, we could have a badass Cougar Character.