Spawners: Add the ability to spawn new creatuers and characters
This is a bit of a big patch, but as is so often the case, things get away from me
a little bit. I've got to start doing things a bit more incrementally. Practice.

Anyway.

This adds the following:
1. A factory that will spawn new animals of type 'rat'. These are pretty simple
   little factory classes, and I'd like to keep them that way so we can add lots
   of them.
2. A Spawner. Spawners are instantiated in rooms and generally owned by a room,
   but are persisted to the database and loaded from it as well. They keep the
   logic for when and how to spawn a new NPC. Right now, the trigger logic is
   pretty simple - after N ticks, they look to see if the monster they are
   tracking is still alive. If not, they make some more of them.

This also updates the world to use all of these things, which is quite a lof of
changes. I expect a bunch more refactoring to follow.