Spawner: Add some tests and fix some bugs
This patch adds tests for the Spawner. As we tested things, it
becamse clear that some of the world mechanisms were a bit odd, as
things that were not using the world singleton were referring to a
different world than those that did. This pushes things towards the
singleton model, which is generally how we expect this to work.