combat: Add a Combat class for determining combat between two characters
This adds a new class, Combat, that manages the combat between an attacking
character and a defending character. When the game loop fires, it should
call the onTick handler for this class, which will process a combat round
for the attacking player trying to hit and damage the defending player.

It will be up to other classes to manage the overall combat between two or
more players.