Add an InanimateModel
This adds a new model for tracking inanimate objects that cannot be worn
nor can they be equipped or wielded as a weapon. They contain some of the
same properties as the existing Armor/Weapons however, and can be placed
in rooms or carried by characters. As a result, we've split out the inanimate
refs into a new sub-schema that Rooms and Characters reference.