Combat: Send combat messages to the entire room
This patch updates combat so that when two characters are fighting,
the combat messages are broadcast to the rest of the participants in
the room. Because our existing `sendImmediate` function in a Room looks
for a single participant to see if we should exclude them, we've updated
the logic in room and Character to suppress messages based on lists. At
the same time, `sendImmediate` is now tolerant to both a list of `Character`
objects *or* a single `Character` being passed in.