Combat: Add hit locations
This patch adds hit locations. Right now, this is just a text description of where
an attacker is targetting and, if it hits, where it hit the defender. Next step is
to add Armor Class.

We also updated the Room:sendImmediate function to not try and differentiate between
an array and a character. Both are objects, and Array.isArray wasn't quite working as
I hoped. Now we just require it to be an array, which is fine.