Combat: Use default weapons on the characters or the weapon they are wielding
This patch does two things:

First, it establishes default attacks on characters. This could be a punch,
a bite, or some combination thereof. Models set this up when created in a
generic fashion that act as a pseudo-lightweight weapon. Combat extracts this
from the player through a new 'attacks' method.

Second, Combat now respects weapons. Because the Combat module doesn't want
to make a lot of these determinations, it asks the Character for their attacks.
The Character makes the determination on what all attacks the Character can make
in that round and returns them.

Note that there's always going to be a bit of a dance between the two, as the
Combat module doesn't want to figure out if the Character is wielding two weapons
(and gets a penalty) or has a versatile weapon and is doing so with two hands.