Character: Fix energy cacluation
If a character had sufficiently high strength, they would actually gain
energy as they moved. This was ... odd. To say the least.

Now, we cap the delta penalty at '1' so that, no matter how strong you
are, you spend a small amount of energy moving. It's enough that you
may need to rest if you go bananas moving around, but a strong character
should generally re-charge quickly. We can come back and drop it to 0
if it isn't fun.