Add Rest and Standing
When you've taken damage, you need to heal up. Currently,
characters don't regen hitpoints without resting. So now we have
a Rest and Tand commands, where standing effectively undoes
resting.

The tricky bit with this is characters now have some state that
they're associated with. We want to prevent resting when we're
fighting, and we want to prevent a character from moving or attacking
when they're resting. We also have to reset state back to normal
when folks are done attacking.

This patch *should* take care of all of that. At the same time,
I don't love state in a variable, as it can get a bit brittle.
And we do have to check it outside of the character in a few places.
So we'll want to be careful with this.