Character: Warn if the room the character is moving to doesn't exist
In some weird cases, we may end up in a situation where the character
tries to move into a room that doesn't exist. Could be the database is
funky; could be some other error. Either way, it's hard to know why
the move doesn't work if we don't log something.