Look: Add some error responses when the action doesn't make sense
If someone says to look in a direction and the direction isn't real,
ask them where they want to look. If someone says to look but it isn't
at something, ask them what they want to look at.

Eventually we'll want to look 'in' things for containers, but that's
a problem for another day.