commands: Introduce an ErrorAction on command failure
Rather than silently swallowing errors, we should return an ErrorAction.
Originally I had thought about handling this in the command sets, but
while many of the errors are similar, some are different enough that it's
probably best to have a bit of duplication and let the commands generate
their own ErrorActions. They were going to have to do it anyway.