RoomModel: Add loading by external identifiers
This is the first patch that adds the ability to create/update
definitions during load. As we use the auto-generated ID quite a lot
everywhere right now, and it's useful to just go with the existing
way that MongoDB handles that ID, we add some additional load properties
that allow us to create and update Rooms with a predictable
external identifier.