Add loading of the game world from JSON
Prior to this patch, the way to make a game world was either through
scripts or by directly creating objects in Mongo. This is kind of a pain,
and easy to end up with references going all which-way. Rather than
continue to try and hand roll areas/rooms and other fairly permanent things,
we can now load the basic game world from JSON. This is relatively forgiving,
in that the game world will only load if the version has incremented, and will
only add/update things it knows about.