SpawnerModel: Support loading spawners from a JSON definition file
While we can load up areas and rooms, we didn't have hooks for spawners.
This patch adds that, along with the expected tests. Since spawners have
a lot of optional attributes - and rooms don't have to have a spawner -
this also adds some additional checks for rooms and/or areas not having
any spawners defined in their JSON or in the JSON file itself.