Spawner: Support passing data from the model to the factory
We want spawners to have the ability to customize their factories.
This is going to be a bit interesting in that we don't really know
what all we want to inject into them just yet, but a good example
is the notion of customizing a basic human character with some
various properties so that we don't have to make a bunch of spawners
just the vary the gender.