Add a Human NPC factory
This patch adds the start of generating human NPCs into the world. Because NPCs are
something that typically has more interactions (quests, shopping, speech, etc.) the
factory to generate them is likely to get a bit more complex than your typical
Rat. So we're calling it out explicitly that it's for NPCs.

Which is probably overkill. Maybe we'll rename it to HumanFactory. Naming continues
To Be A Thing.