Clean up test leaks from World shutdown
If you have a World, you have to shut it down. Otherwise the event loop will
just continue tick on and on.

This is a bit tricky at times to catch, as we use a singleton to get the
World a Room is in. This pattern may eventually backfire and we'll need to
be a bit more deliberate with how we get a handle to the World, but for now
we'll be okay with this.