Refactor InanimateContainer to an ObjectContainer and use it for Characters
When there are multiple characters in the same room, we need to have some
way for text commands to differentiate between the 'things' that are there.
Luckily we are already using that for inanimates, and because we don't care
about what we're putting in that underlying container so long as it has a 'name'
field, we can just reuse it.