Create a Door Model
After some internal debate, I've decided that Doors really are their
own thing and shouldn't be bundled in with Inanimates. What pushed
me in this direction is that doors have state - they're open or closed,
they have lock properties, etc. We have to keep track of that state
somewhere, and none of it belongs with other inanimate objects.
Arguably, containers are a special type of inanimate and we bundled
that into the same model as everything else, but at least that's just
an array and not something that has lots of different and intersecting
state properties.

So, Doors are a thing. Huzzah.