BaseClass: Prevent indexing errors for Level 10 classes
At times, we want to show how much experience is needed to get to the
next level. That means there should always be a next level, or else
we'll blow past the entries in the class to experience mapping. There's
probably some better ways to handle this (such as wrapping this up in
a tighter function), but for now putting a "Level 11" in there with
an obscene amount of experience required works.