Create detailed character views
When examinine a character - or whenever we want to dump out a long
text description of them - we'll want a lot of information. This is
a function of both the skill check of the observer as well as the
ability of the character being observed to hide things. Since this
is pretty complicated and lengthy, I've moved this into a new 'helpers'
folder for Characters. I'll probably move more stuff there over time
to keep the Character class from getting too bloated.

For now, this does an okay-ish job of keeping a lot of text formatting
along with a lot of skill checks out of the basic character class.
This was written with the notion that as you get better at observing
things, you'll see more. And if you're observing yourself, via whatever
command, you'll get all the details on yourself.