Examine: Prevent examining yourself
Just like Look doesn't allow you to look at yourself, Examine should
not either. Examine is about a detailed study of a person or thing,
and for now we should keep it focused on the notion of looking intently
at things.

The actual output of an examination of yourself should be a game command
like 'score'.