Look: Allow looking at doors
This allows the player to start looking at doors, in the same way that
they can look at objects in a room. It sends back a detailed description
of the door, without going into their properties (which is more of an
examine kind of thing).

While we have to do some parsing of the 'direction.{name}' in order to
find a door when there are multiples in a room, I haven't yet decided
if that warrants its own container. We'll definitely need to do that
parsing again for other times when you want to interact with a door
on a particular location, which _is_ going to happen on at least 5 other
actions. So we'll want to stick it somewhere, but it's really all about
just finding the thing in the room you want to interact with. So maybe
just a convenience function at some point.