Refactor looking up a door
In preparation for a lot more stuff trying to find if there's a damn
door in this room, we move this into a common routine, 'getDoor' on
a Room. This does change a tiny bit the messages that are sent back
for Examine and Look, but not enough that I care.