Add Factions as a first class model / definition
Factions are one way in which characters will have a reputation, and, arguably,
deal with the consequences of their actions. To start, we need a model that
can store that a faction exists, and what factions are opposed to that faction
or support that faction. Furthermore, some factions may be more inclined
to look favorably on actions that advance their needs, while other factions may
penalize you for more negative actions. Anyway.

FACTIONS.