Make it easy to generate objects with a factory map
Right now, you have to import a factory everywhere you want to use it to
construct an object. That's annoying.

This creates a centralized map of factories that can be invoked from some
other piece of code. Because we have a lot of factories, we now have some
generated unit tests that make sure that they can generate objects given
a variety of input scenarios.