Support generation of Human Fighters
We can now specify a human fighter. Their definition can include the
level of their class, and we'll set their level correctly and update
their skills to max out at that level - which is appropriate for an
NPC. We can give them equipment, and we'll generate the equipment
and put it in those slots.

This won't work for other class types just yet as we haven't really
defined what happens to those classes when you go past level 1.