Add Conversations
This is a pretty hefty first addition for conversations, but it's a nice 'start'.
This adds a state machine that manages the conversation on a player by player
basis. While it really only implemented 'onSay', it provides the start of hooks
for what characters should say when attacked or when they die. There's a lot more
to work on here: faction statuses, not starting a conversation when you're being
attacked or in combat, etc. - but it gets the basics going.

Note that we also need to add 'quests' as a thing, and have a hook for that in
sept.json.