Character: Add conversations
A character can have a conversation associated with them. If they do, it will
reference an ID for the conversation model. The `load` and `save` routines
have properly been updated to load and save the Conversation, although as you
would expect, the `Conversation` class itself manages those calls. If a conversation
exists on the Character, then it will be invoked with whatever is said in a Room.