CurrencyManager: Add a basic currency manager for characters
There's a bit of special logic we want to apply when tracking a character's
money. This class wraps up that logic so that we can safely track currencies,
and 'withdraw' it in a reasonable way that doesn't result in negative numbers.
We can theoretically add exchanges at some point as well.