Add currency to characters
A character can now walk around with some money in their pocket. Money
is treated differently from other Inanimate objects, in that a character
does not hold it in their hauled inventory but instead stores it as a property
on their character via the CurrencyManager. However, if the character dies
we turn it into an Inanimate object that can be picked up and manipulated.

Subsequent patches will do the bit that turns the Inanimate object into
cold hard cash on the Character.