ObjectContainer: Support money look-ups
When we create an Inanimate that represents money, we add a qualifier
to the name, e.g., "gold (50)", that tells the player how much gold
is in the pile. The (50) is for informational purposes only. As such,
when doing look-ups, we want to scrape out and remove that (50) from
the name look-ups.

In hindsight, this probably should have just been part of the Inanimate
toShortText() method. Oh well.