QuestModel: Add a backing model for Quests
This is *mostly* complete, as it will allow us to create a Quest
and manipulate it. The Model at this point does not yet really store
character progression, but I'm holding off on adding that until I figure
out where and how I want Quests to be manipulated in the world.

They're kind of a weird shared resource, which makes me think I'll stick
them in the World.