Update characters for the Rat Quest
As we add in the first quest - KILL ALL THE RATS - we need to put
some support into the RatFactory to know that all of these rats it
is generating are the same rat that we need the character to kill N
of. That means we need a unifying 'tag' for the rat, which we provide
via a new field 'characterRef'.