Character: Load/store factions
Factions aren't currently loaded or stored on a character. We now
properly do that. This includes adding an 'initialize faction'
function which will set an initial faction value if one doesn't
already exist.

There's a bit of a race here that's potentially possible that we'll
have to think about. Creating new player characters will need to have
some default factions applied, which is a one time thing. Normally
we could stick that in the constructor of the PlayerCharacter class
and then have the load() function overwrite it, but that probably
won't work here since I made initializeFactions a NoOp if a faction
score already exists.

What this really means is that we may be best off creating a 'default kit'
for PlayerCharacters that is applied when the character is first created
through the REST API. That's a good 'set it and forget it' place, and
won't run afoul of load order issues. (It would also keep the constructor
clean)