Quests: Add toText and toJson methods
(There's probably a few that crept into the previous patch. Note to self:
get in the habit of committing way more often)

Add some ability to convert quests into text or JSON. For JSON, this is
needed when we store the quests (which means that should be in the previous
commit, drat.) For text, we'll need it for our commands.