Character: Process queued actions for Combat and onTick
This patch adds a queue of actions on a Character. We're likely to
refactor this at some point, but for now we keep a single queue of
actions that can be special attacks, as well as 'other things' that
tick down in the onTick handler.

We also don't regenerate in combat. I may tweak this again later, but
for now having HP/mana/energy stop regening while you're fighting for
your life is a good reason to make characters pause and rest.