Kick: Add the ability to kick people
Fighters like to kick. So now they can. When you kick (via a new
command), the character queues up a kick attack for the next combat
round. When the combat round happens, it dequeues the special attack
and kicks the target.

Combat can also be started by kicking someone. If someone initiates
with a kick, it will start a combat round on the next tick.