Add an ActionEffectQueue for handling attack actions and character effects
Both actions and effects have a timer on them that slowly ticks down until
they fire. This creates a queue for both that is processed in attacks and
the onTick handler. Items in them 'burn down' until their tick property is
0, at which point they're returned. In the case of attacks, that means they
are performed; in the case of effects, that generally means they stop
applying to the character.