DiceBag: Add the ability to set the next result
We already have the need in Combat to override dice rolls during
tests. Turns out we need to do that a lot. This adds the ability
to do that on any DiceBag. This is needed for tests, but we also
will need to do that for certain skills/feats/etc.