AssassinationQuestStrategy: Handle text substitution when the count is 0
When we haven't killed anything yet, there is no state showing how many
characters we've killed. As a result, the substitution will skip over
replacing the text.

Rather than do the check to see if we've killed anything yet, we instead
just use an '|| 0' to handle the case where the result is undefined.