Add an Admin command set and a 'Kill' command
We're definitely in test mode. That means sometimes, we might need
to nuke a character in the game from high orbit. This patch adds an
admin command set for, well, administrator commands, and a 'god mode'
kill command. Because we don't want this to ever get aliased, it's
named 'admin-kill', which also sets a pattern for future admin commands.

Because I don't have a way to say who is an admin and who is not (although
we should add that at some point), for now all PCs get the admin command
set. We can futz with this later.