Character: Unsubscribe from the message bus on death
We're currently orphaning the subscription to the message bus when
a chaaracter dies. That means that the Character object itself is
still going to linger, as there's a reference to the character via
the subscription to the message bus. EEeeeekkkk.

This now unsubscribes the character when they die. We could get more
aggressive here and if we receive a message while dead force the
unsubscription... but I'm going to hold off on that. We really
only subscribe to rooms, so that makes it easy to know what to remove
for now.