Conversation: Fix characters triggering conversation without speaking
We should only trigger a conversation when a character says something.
We can only tell that when the `packet`'s `message` has a `socialType`
of `say`, which is not currently plumbed through to the Conversation
class.

So we do that. This is a relatively small change but did require some
refactoring of the tests.

closes: #66