Party: Add parties as a first class object in the game
Parties are collections of Characters broken into two lists: invitees
and members. The size of a party cannot exceed some fixed bound, where
the size is the number of invitees + the number of members. As members
join, they are moved from the invitee list to the members list if
present in the invitee list. A single member is set as the leader of
the party, and a registry tracks the created parties tied to specific
leaders. Parties are persisted/loaded to the MongoDB database.

Finally, a small factory function has been provided to make it easy to
create parties given a party leader.