Party: Add the ability to accept invites
Once you're invited to a party - or multiple parties - you need to
accept the invitation. This is done by using the 'party accept' command,
which requires you to specify which party to accept. Upon accepting it,
you'll be put in a party with the leader of that party, and any pending
invites you had to other parties will be cancelled. The party leaders of
the other parties are notified when this happens.