Party: Add experience for party members
If you're in a party, you now get a small experience bonus based on
the number of people in your party. You also, however, split experience
across all party members. This helps to make large parties (generally)
a good thing in the long run, but fair so that we don't have parties
levelling up 4x faster.