Party: Add the metadata handling for Auto-Follow
Nothing makes use of this yet, but we want to be able to follow a character in
a party. Since we're going to have lot of behavior tweaks in parties, it makes sense
to capture them in dedicated handlers. The handlers are responsible for parsing the
metadata and managing their state based on the changes that are triggered.

This is nice, but because party metadata requires a lot of handling of in-game
state - in this case, the 'move' event from characters (TBD), we don't save this
to the database. Maybe someday that will be worth doing, but right now it feels
okay to make this transitory with logging in/logging out/restarting the server.