Party: Add a container for party metadata
When you're partying and you want to know who to follow to the cab,
it's time for metadata. This adds a container wrapper for tracking
individual metadata that has been set for specific characters. There's
a hook to add future metadata handlers, similar to auto-follow, and we've
wired in auto-follow.