Party: Add tracking of metadata
Now that we have a metadata container and the ability to auto-follow (the
first implementation of metadata), we need to wire it into the Party. This
does the bare minimum of that, and also modifies the toJson method to dump
the current metadata out for a character as a list (in case we want to display
it somewhere).