CombatManager: Prevent bug where second attacking character is stranded
Imagine the following combat:
 A attacks B
 C attacks B

In the case where A kills B, we would remove the combat for A and B and
would do a reverse lookup at the time of the death to find the combat for
B and A and remove it as well. That reverse lookup would also remove any
combats (if they existed, which they wouldn't) for B and C. But we didn't
then take C and remove the combat for them and B, which would leave C
fighting thin air.

This patch alters that so that we now look-up and cache all combats going
to B when B dies. We then remove those combats.