PartyMetadataAutoAttack: Add the ability to auto-attack for someone
When you're in a party, a lot of utility characters (which is sort of everyone
that is not a Fighter, and maybe a Rogue) probably just want to be managed
and every now and again cast a spell. Still, taking a free swing at an enemy
can be handy. Auto-attack to the rescue.

You can now set 'auto-attack on <party member>' in your party, and you'll
auto-attack whoever they attack.