Room: Create dark rooms
When a room is too dark to see, we will now send a room description
that notes that the player cannot see anything. This is tracked as a
permanent attribute of a room on its model. Because we may end up wanting
a lot of other attributes on rooms - such as 'sky' (with gravity!) or a
room filled with some poisonous gas - this is a list of attributes that
are maintained on the room.