ObjectContainer: Add updates for handling a room
When an object container is part of a room, properties on the room may
change how things interact with the container. Rather than adding a silly
number of methods on the Room class, we instead add the Room as a property
on the container that changes how find/filter/findItem work.

Note that we also now have a find/filter option. These should be used by thing
wanting that room interaction behavior instead of snagging the underlying
container via `all`.