Create explicit AI for wandering around
This starts to extract a character's AI into something more manageable
than where we had it. Now, AI can be loaded in and saved on a Character
model. Because we don't want to accidentally start triggering AI for
a Player Character, we now have another class, NonPlayableCharacter.
This creates an interesting dilemma: we want Humans to be both playable
and non-playable characters.

To deal with this inheritance mess, PlayerCharacter and NonPlayableCharacter
are 'mix-in' classes that 'wrap' an actual class that inherits from Character.
Wrapping that class creates an 'on the fly' class that inherits from the
provided base class, letting us add overrides to the default implementations
of methods. This gives us the opportunity to insert AI wrappers in things
like onTick that can't get invoked for a playable character.